gh-secrets/requirements.txt
requests
pynacl
fire
python-dotenv